fix(2-get..) : type change for bit

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -10,7 +10,7 @@
 
 int get_bit(unsigned long int n, unsigned int index)
 {
-	int bit = (n >> index);
+	unsigned long int bit = (n >> index);
 
 	if (n == 0 && index == 0)
 		return (0);
